Add NMake Makefiles
This adds a set of NMake Makefiles that can be used to build libxml++,
which will replace the Visual Studio 2010(!) projects that reside in the
repository (which is probably broken and we actually need Visual Studio
2017+ to build the current code due to a bug in the Visual Studio 2015
compiler).

The unfortunate fact is that we do need to link to a glibmm that is
built with the same compiler, otherwise linking will fail and the
binaries of the dependent glibmm built with other compilers won't work
well with the builds here if they are built with other Visual Studio
compilers, so we will need to separate the output filenames
for the resulting .lib and .dll files for Visual Studio 2015, 2017 and 2019.
A NMake Makefile build options "USE_COMPAT_LIBS" is added if linking to
xxxx-vc140-y_z.lib for glibmm is still desired, such as in the case where
glibmm was actually built with Visual Studio 2017 or 2019.